# docs/dev/web/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Web API
  version: 0.2.0
  description: |
    Current Atlas Hub API surface from route handlers in apps/web/app/api and apps/web/app/download.
servers:
  - url: https://atlas.example.com
    description: Replace with your Atlas Hub deployment URL
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OidcHeader:
      type: apiKey
      in: header
      name: x-atlas-oidc-token
...
paths:
...
  /api/v1/packs:
    get:
      summary: List packs visible to current user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Pack list
    post:
      summary: Create pack
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Pack created
...
  /api/v1/releases/{product}/{version}/{os}/{arch}:
    parameters:
      - name: product
        in: path
        required: true
        schema:
          type: string
          enum: [launcher, cli, runner, runnerd]
      - name: version
        in: path
        required: true
        schema:
          type: string
      - name: os
        in: path
        required: true
        schema:
          type: string
          enum: [windows, macos, linux]
      - name: arch
        in: path
        required: true
        schema:
          type: string
          enum: [x64, arm64]
    get:
      summary: Resolve specific distribution release metadata
      responses:
        '200':
          description: Release metadata with asset download IDs
        '404':
          description: Release not found
...
  /api/v1/download/{downloadId}:
    parameters:
      - name: downloadId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Universal download resolver for immutable artifacts
      responses:
        '302':
          description: Redirect to object storage URL
        '404':
          description: Download ID not found
...
  /api/v1/launcher/updates/{channel}/{os}/{arch}:
    parameters:
      - name: channel
        in: path
        required: true
        schema:
          type: string
          enum: [stable, beta, dev]
      - name: os
        in: path
        required: true
        schema:
          type: string
          enum: [windows, macos, linux]
      - name: arch
        in: path
        required: true
        schema:
          type: string
          enum: [x64, arm64]
    get:
      summary: Channelized launcher updater view from release registry
      responses:
        '200':
          description: Tauri updater response
        '404':
          description: Updater release not found
...